Add Views, rearrange things
