+++ app/javascript/components/modals/track-welcome-modal/LHS/BootcampRecommendationView.tsx
@@ -2,12 +2,21 @@ import React, { useContext } from 'react'
 import { TrackContext } from '../TrackWelcomeModal'
 
 export function BootcampRecommendationView() {
-  const { hideBootcampRecommendationView } = useContext(TrackContext)
+  const { hideBootcampRecommendationView, track } = useContext(TrackContext)
   return (
-    <div className="grid grid-cols-2 gap-12 items-center">
-      <h2>You might find the Bootcamp is a better fit</h2>
+    <>
+      <h4 className="text-h4">You might find the Bootcamp is a better fit</h4>
 
-      <p>Here's why:</p>
+      {/* TODO: Add copy */}
+      <p className="mb-16">
+        Here's why:
+        <br />
+        <ul className="flex flex-col gap-4 text-14 font-medium">
+          <li>✅ You're new to {track.title}</li>
+          <li>✅ You want a structured learning path</li>
+          <li>✅ You want to build projects</li>
+        </ul>
+      </p>
 
       <div className="grid grid-cols-2 gap-12 items-center">
         <a className="btn-m btn-primary">Go to bootcamp</a>
@@ -18,6 +27,6 @@ export function BootcampRecommendationView() {
           Continue anyway
         </button>
       </div>
-    </div>
+    </>
   )
 }
