+++ app/javascript/components/modals/track-welcome-modal/RHS/TrackWelcomeModalRHS.tsx
@@ -1,0 +1,18 @@
+import React, { useContext } from 'react'
+import { TrackContext } from '../TrackWelcomeModal'
+import { VideoRHS } from './VideoRHS'
+import { WhoIsThisTrackForRHS } from './WhoIsThisTrackForRHS'
+
+export function TrackWelcomeModalRHS(): JSX.Element {
+  const { track, currentState, shouldShowBootcampRecommendationView } =
+    useContext(TrackContext)
+
+  if (
+    currentState.matches('learningEnvironmentSelector') &&
+    shouldShowBootcampRecommendationView
+  ) {
+    return <WhoIsThisTrackForRHS track={track} />
+  }
+
+  return <VideoRHS track={track} />
+}
